homework6/pyproject.toml
[tool.poetry]
name = "homework4"
version = "1.0.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
rq = "^1.10.1"
redis-server = "^6.0.9"
uvicorn = "^0.17.6"
fastapi = "^0.75.1"
python-multipart = "^0.0.5"
redis = "<=4.2.0"
Pillow = "^9.1.0"
types-redis = "^4.1.19"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pytest-cov = "^3.0.0"
autoflake = "^1.4"
isort = "^5.10.0"
black = "^22.1"
unify = "^0.5"
flake8 = "^4.0"
pylint = "^2.12.1"
mypy = "^0.940"
pytest-deadfixtures = "^2.2.1"
fakeredis = "^1.7.1"
mock = "^4.0.3"
types-mock = "^4.0.13"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
